requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
black==25.1.0
bleach==6.1.0
boto3==1.39.9
botocore==1.39.9
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.9.2
cryptography==45.0.5
daphne==4.0.0
distlib==0.4.0
dj-database-url==3.0.1
Django==5.0.14
django-allauth==65.10.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
factory-boy==3.3.0
Faker==37.4.2
filelock==3.18.0
firebase-admin==6.5.0
flake8==7.3.0
freezegun==1.4.0
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
httplib2==0.22.0
hyperlink==21.0.0
identify==2.6.12
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
mccabe==0.7.0
msgpack==1.1.1
multidict==6.6.3
mypy==1.8.0
mypy_extensions==1.1.0
nodeenv==1.9.1
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre-commit==3.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==6.0.0
psycopg==3.2.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyotp==2.9.0
pyparsing==3.2.3
pypng==0.20220715.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
sentry-sdk==1.40.6
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.3.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.33.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.20.1
zope.interface==7.2
